refactor: delete waste functions
