+++ src/const.js
@@ -17,27 +17,26 @@ const SortType = {
 const DEFAULT_SORT_TYPE = SortType.DAY;
 
 const DateTimeFormat = {
+  TIME: 'HH:mm',
   SHORT_DATE: 'MMM D',
   EDIT: 'YY/MM/DD HH:mm',
   DATETIME_DAY: 'YYYY-MM-DD',
   DATETIME_FULL: 'YYYY-MM-DDTHH:mm',
-  TIME: 'HH:mm'
 };
 
-const WAYPOINT_COUNT = 4;
 
 const UserAction = {
-  UPDATE: 'UPDATE',
   ADD: 'ADD',
-  DELETE: 'DELETE',
+  UPDATE: 'UPDATE',
+  DELETE: 'DELETE'
 };
 
 const UpdateType = {
   PATCH: 'PATCH',
   MINOR: 'MINOR',
   MAJOR: 'MAJOR',
-  INIT: 'INIT',
-  ERROR: 'ERROR'
+  ERROR: 'ERROR',
+  INIT: 'INIT'
 };
 
 const Mode = {
@@ -54,8 +53,8 @@ const Method = {
 };
 
 const URL = {
-  WAYPOINTS: 'points',
   DESTINATIONS: 'destinations',
+  WAYPOINTS: 'points',
   OFFERS: 'offers'
 };
 
@@ -73,7 +72,6 @@ export {
   FilterType, DEFAULT_FILTER,
   SortType, DEFAULT_SORT_TYPE,
   DateTimeFormat,
-  WAYPOINT_COUNT,
   UserAction, UpdateType,
   Mode,
   Method, URL,
