+++ src/view/edit/edit-view.js
@@ -1,7 +1,7 @@
 import AbstractStatefulView from '/src/framework/view/abstract-stateful-view';
 import createEventDetailsTemplate from './template/event-details';
 import createEventHeaderTemplate from './template/event-header';
-import { getObj, isValidDateInterval } from '/src/util/util';
+import { getObj, isValidDateInterval } from '/src/util/common';
 import { Mode } from '/src/const';
 import flatpickr from 'flatpickr';
 import '/node_modules/flatpickr/dist/flatpickr.min.css';
